add reasonable defaults
