+++ test-utils/test-runner/src/lib.rs
@@ -242,6 +242,7 @@ mod host_functions;
 
 pub use host_functions::*;
 pub use node::*;
+pub use utils::*;
 
 /// Wrapper trait for concrete type required by this testing framework.
 pub trait ChainInfo: Sized {
